+++ dashboard/src/contexts/account/index.ts
@@ -1,0 +1,3 @@
+export * from './account-provider';
+
+export * from './use-account-context';